Add files via upload
I tried to upload to git through my terminal but it did not cooperate even though I was putting in the correct password it kept giving me this error: fatal: Authentication failed for 'https://github.com/oobasuyi314/4232Capstone/' so I just put the game file in the repository through the git website. if it doesn't load the code please just email me instead of taking points because I can just send you the game folder with all material in it (I will just go ahead and send it anyways to Professor Jugan).